requirements.txt
PySide6>=6.5.0
Pillow>=10.0.0
google-generativeai>=0.3.0
mss>=9.0.1
pynput>=1.7.6
python-dotenv>=1.0.0
numpy==1.26.4 